Python_Programming_Blueprints/musicterminal/requirements.txt
requests==2.18.4
PyYAML==3.12
flask==1.1.1